Feature/neural network implementation

+++ include/normal_distribution.h
@@ -8,6 +8,6 @@ void seed_random(void);
  * @brief Generate random number from normal distribution N~(0,1)
  * Using the Box-Muller transform
  */
-float rand_norm(void);
+double rand_norm(void);
 
 #endif
